Improve robustness of code ...
to handle the cases when both prefetch *and* normal requests fail.

(e.g. in Safari 8 and 9 for <link rel=prefetch> tests)

This led to the situation where `test.resource` did not exist at
end-of-test and so the server code trying to modify `test.resource`
failed.

ResultsTable.js now also takes more care to validate/check-existence-of
the data before using it.